src/main/java/com/EventHorizon/EventHorizon/RepositoryServices/Mappers/DetailedLaunchedEventDtoMapper.java
package com.EventHorizon.EventHorizon.RepositoryServices.Mappers;

import com.EventHorizon.EventHorizon.DTOs.EventDto.AdsOptionDto;
import com.EventHorizon.EventHorizon.DTOs.EventDto.DetailedLaunchedEventDto;
import com.EventHorizon.EventHorizon.DTOs.UserDto.OrganizerHeaderDto;
import com.EventHorizon.EventHorizon.Entities.EventEntities.Event;
import com.EventHorizon.EventHorizon.Entities.EventEntities.LaunchedEvent;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class DetailedLaunchedEventDtoMapper {
    @Autowired
    private AdsOptionDtoMapper adsOptionDtoMapper;

    public LaunchedEvent getEventFromDetailedEventDTO(DetailedLaunchedEventDto dto) {
        Event event = Event.builder()
                .name(dto.getName())
                .description(dto.getDescription())
                .eventCategory(dto.getEventCategory())
                .eventDate(dto.getEventDate())
                .eventAds(this.adsOptionDtoMapper.getAdsOptionFromDTO(dto.getEventAds()))
                .eventLocation(dto.getEventLocation())
                .build();
        return LaunchedEvent.builder().event(event).launchedDate(dto.getLaunchedDate()).id(dto.getId()).build();
    }

    public DetailedLaunchedEventDto getDTOfromDetailedEvent(LaunchedEvent event) {
        DetailedLaunchedEventDto detailedLaunchedEventDto = DetailedLaunchedEventDto.builder()
                .id(event.getId())
                .eventAds(new AdsOptionDto(event.getEventAds()))
                .eventOrganizer(new OrganizerHeaderDto(event.getEventOrganizer()))
                .eventDate(event.getEventDate())
                .name(event.getName())
                .launchedDate(event.getLaunchedDate())
                .eventLocation(event.getEventLocation())
                .eventCategory(event.getEventCategory())
                .description(event.getDescription())
                .build();
        return detailedLaunchedEventDto;
    }
}
